KURSP-805: Adapt workflows to build and push 'dist' to azure
Change the workflows to produce cleaner commits with removing and
ignoring the dist folder. The workflows will now handle building and
distributing the dist folder contents to azure blob storage, whereas
before it only distributed it. The former solution could produce
scenarios where the source code and the dist folder content was out
of sync thus pushing out of date code to azure.

    - Replaced 'deployToMaster.yml' with 'build-master-push-azure-yml'
      Replaced 'main.yml' with 'build-stage-push-azure.yml'.
      This gives the workflows an apt name for what they achieve.
    - Modified '.gitignore' to ignore 'dist' folder ensuring
      local builds do not polute remote.
    - Remove 'dist' folder from remote repo.
    - Workflow builds 'dist' and push it to azure. For master branch
      this blob storage is called 'udirdesign' and for the staging
      branch it is called 'udirdesign-staging'