KURSP-1064: Improve loadspeed of course module tree (left-side)
Increases the speed at which the left course module tree loads.
This is done by using vuex to keep the module state in the browsers
session storage. Session storage lasts in current window in current tab.
The implications of entertaining this state in the browser means that
other features needed to be changed to keep sync between vuex state and
server state. Such as clicking on a link in the module browser causing
the setting of the active page. 'Forrige' and 'neste' button needed to
be changed for the same reason. ModulesProgressIndicator needed to be
updated and button for 'Ferdig' needed to be changed. Taking account of
users disallowing session storage and server changes to the model
by creators was needed for a fluent experience.
Server version is the correct version, thus server version is to
replace web browser version given a difference.

    - Add vuex store, after first load module tree is stored.
    - Changes in server model replaces vuex model.
    - Not allwoing sessionstorage causes default behavior
    - Add wiring to keep state in vuex and server to be in sync
    - Add extra reactivity, progression bar and 'Ferdig' button
      now updates before page is changed.
    - Improved load speed of course module tree, 'forrige', 'neste'
      button, progression bar and checkmark for done.